refactor: move types to types file
